refactor(middleware): move container.ts into agent/middleware/, add shutdown tests
- Move sandbox/container.ts → agent/middleware/container.ts
- Fixes reverse dep (sandbox → agent/middleware) and the mid-file ExecResult
  import in container.ts; dependency now flows agent/middleware → sandbox/config
- sandbox/ now contains only config types
- Add 3 tests for AgentRuntime.shutdown: no-op without sandbox, idempotent,
  cleanup-once across repeated calls
