+++ src/agent/middleware/executor.test.ts
@@ -2,7 +2,7 @@
 
 import { describe, it, expect, vi, beforeEach } from "vitest";
 import { SandboxExecutor } from "./executor.js";
-import type { ContainerManager, ContainerInfo } from "../../sandbox/container.js";
+import type { ContainerManager, ContainerInfo } from "./container.js";
 import type { SandboxConfig } from "../../sandbox/config.js";
 
 // ---------------------------------------------------------------------------
